Rename ScreenClient to Screen and move all screen state handling there (bug 1597418)
Previously, it handled tracking screen state for the primary display and ContentsView contained its own tracking of screen state for the display it was viewed on. In addition to this, the tracking of screen state was scattered around all over, making it difficult to test.

This change moves all of this in to one class. It makes it easier to test, and means that there is now only 1 mechanism to test.