Refactor WebPreferences ownership
This makes oxide::WebPreferences reference counted, which allows us to get rid of WebPreferencesObserver::WebPreferencesDestroyed() and generally simplify how WebPreference creation is handled. A consequence of this is that WebPreferencesObserver goes away entirely, replaced by callbacks instead.

This also adds WebContentsClient, which I was initially going to use in this. However, it ended up being unused but I've kept it around as it's going to be used in other work.