Initial implementation of context menus for UbuntuWebView (LP: #1637186)
This adds support for both context menu styles currently used in webbrowser-app. There currently aren't any
tests for this - the code is not enabled by default anywhere, nor is it ready to use. Tests are being tracked
in a separate bug.

For the browser to start using this, UbuntuWebView needs an API to allow the browser to customize the menu
contents.