Fix a test hang at the start of tst_focus.qml (LP: #1656303)
tst_focus.qml hangs if it runs after tst_bug1626099.qml, which used to be the last test in this sequence. tst_bug1626099.qml spins the CPU in its unload handler, and to avoid this being optimized out, it logs a console message on each loop iteration. It looks like this was flooding the browser process's IO thread, causing problems for subsequent tests.

This updates the test to avoid IPC on each loop iteration, and also ensure the webview is deleted synchronously at the end of the test.