Change the behaviour of JS dialogs (LP: #1568296, LP: #1656905)
This commit makes the following changes to the way that JS dialogs work:

- Alert dialogs (window.alert()) requested from background webviews are
no longer displayed immediately, but will instead be displayed when the
webview is focused. To prevent a background webview from blocking webviews
with which it might share a process, window.alert() will return and script
execution will continue immediately in this case.
- Confirm (window.confirm()) and prompt (window.prompt()) dialogs
requested from background webviews are now suppressed.
- Alert, confirm or prompt dialogs are now dismissed automatically if
another webview is focused.
- beforeunload dialogs associated with navigations that are not
application initiated now require a user gesture. If there is no user
gesture then the navigation will proceed without a prompt.

* Note, the foreground webview is the currently focused webview or the
last one to have active focus.

There are some issues with these changes:

- The Ubuntu UI Toolkit dialogs are basically window modal. Alert dialogs
can be missed if one is displayed in an application that is currently not
focused, and the user subsequently focuses a webview in another window
from that application (as this will result in the dialog being dismissed).
I think the fix for this is probably to block input to other application
windows until the window with the dialog has been focused.
- Related to the last point, a beforeunload dialog displayed in one window
can block webviews in another window if they share content processes, but
this isn't a new issue. The fix here is probably to block input to other
application windows until the beforeunload dialog has been dismissed.
- There is no indication that a background webview has called
window.alert() until the webview is focused. Ultimately we need another
API for this, and then webbrowser-app will consume this to either request
attention to the originating background tab or automatically focus it
(which is what Chrome does).