Add touch editing UI to UbuntuWebView (LP: #1638852)
This is an initial implementation of the touch editing UI for
UbuntuWebView. It currently lacks provision for embedders to customize it -
this will be added later.

Although there are extensive unit tests for TouchEditingMenuControllerImpl,
there are currently no higher level QML tests, as the QML test suite has no
API for simulating touch events. These need to be added later